+++ Quotation/Infrastructure/Quotation.DataAccess/Models/OwnerDetail.cs
@@ -21,6 +21,6 @@ namespace Quotation.DataAccess.Models
         public Nullable<System.DateTime> LastUpdatedDate { get; set; }
         public string Email { get; set; }
         public string Address { get; set; }
-        public Nullable<short> RenewalRemindDays { get; set; }
+        public Nullable<int> RenewalRemindDays { get; set; }
     }
 }
